packages/ai-tools/pyproject.toml
[tool.poetry]
name = "ai-tools"
version = "0.1.0"
description = ""
authors = ["Wojciech Majewski <wojciech.majewski@pm.me>"]
readme = "README.md"
packages = [{include = "ai_tools"}]

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
sentence-transformers = {version = "^2.2.2"}
langchain = "^0.1.0"
jsonlines = "^4.0.0"
unstructured = {version = "^0.11.8", extras = ["local-inference"]}
pypandoc = "^1.12"
pyright = "^1.1.344"
rich = "^13.7.0"

[tool.poetry.extras]
embeddings = ["sentence-transformers"]
import_epubs = ["unstructured", "pypandoc"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
